requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
frozenlist==1.3.0
future==0.18.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
pycodestyle==2.8.0
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
